requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
dj-database-url==3.1.2
Django==4.2.29
django-bootstrap-v5==1.0.11
et_xmlfile==2.0.0
gunicorn==25.1.0
openpyxl==3.1.5
packaging==26.0
pillow==12.2.0
psycopg2-binary==2.9.11
python-decouple==3.8
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.12.0
